Swap order of version and release arguments.
We don't want to tag submissions in the repository because they aren't
anyone's business (so there!), so instead of tagging, we'll drop an
XML file into the submitted project telling us how to check out the
version that we submitted (URI + revision).

This means the working copy must consist of a single revision, so we
enforce that.


git-svn-id: https://svn.calendarserver.org/repository/calendarserver@211 e27351fd-9f3e-4f54-a53b-843176b1656c