Add an enabledForCalendaring option (default=True) on IDirectoryRecord which determines whether a calendar home should be provisioned for the record.
In the OpenDirectory case (this should be extended to other
directories as well), load up all groups, but set
enabledForCalendaring to False (and force calendarUserAddresses to be
an empty list) if the group doesn't have a ServicesLocator from this
server.

Rename DirectoryPrincipalResource to
DirectoryCalendarPrincipalResource and create a new superclass
DirectoryPrincipalResource which is not a CalDAV principal, but is
instead a generic WebDAV principal (which doesn't have all of the
CalDAV live properties).  Records that do not have calendar homes are
provisioned with generic WebDAV principals, while those that do get
CalDAV principals.


git-svn-id: https://svn.calendarserver.org/repository/calendarserver@1889 e27351fd-9f3e-4f54-a53b-843176b1656c