Merge my 'sendfdport' branch.
This expands on twext.python.sendfd to add a twisted-style Port object that
listens for incoming connections and dispatches them to IProtocolFactory
providers in a different process using sendmsg().  It also uses that mechanism
to dispatch incoming connections to subprocesses by default, although it can
still be disabled by the 'UseMetaFD' option, for the time being.


git-svn-id: https://svn.calendarserver.org/repository/calendarserver@5425 e27351fd-9f3e-4f54-a53b-843176b1656c