Trace the master *and* the slaves
Tracing only the master means we miss out on any extra I/O the slave
processes do, which is very significant in the filesystem-backend
configuration and potentially for other reasons.  This approach also
works with pre-connectionpooling versions of the code, so we could
regenerate all the stats in this configuration to ensure
comparability.


git-svn-id: https://svn.calendarserver.org/repository/calendarserver@6560 e27351fd-9f3e-4f54-a53b-843176b1656c