Add a "user profile" which invites uses to events.
All clients started up by population.py are now given an "inviter" profile which (rapidly) adds invitees to a random event that it finds.  Still to do is to add an "event creator" profile so that the inviter can find some events, as well as to allow tuning of the inviters parameters instead of hard-coding the loop interval and such.

The inviter also never un-invites anyone from an event, which would probably be a good thing to do.

Clients will also never accept or reject an invitation.  A profile for that should be added as well.

The SnowLeopard iCal simulator also now has a method for adding an attendee to an event which implements the same behavior as the SnowLeopard iCal client.

The httpclient StringProducer is also changed to reject obviously bogus data (eg unicode).


git-svn-id: https://svn.calendarserver.org/repository/calendarserver@6797 e27351fd-9f3e-4f54-a53b-843176b1656c